Correct the definition of the smoothing length, so that the density estimator is that used in Price 2011, eq. 6.
Before I was using a slightly different scaling for the SPH kernel,
which came with a different normalisation. The different for Arepo is
small (because I was internally consistent with this definition) but the
difference for Gadget was large, because the smoothing length was not
what I thought it was!

It differed by a factor of 2 * (4 pi/3)^1/3 ~ pi