Adjust the optical depth to match the new SPH kernel.
This fixes a bug introduced in 4018b6d6fdb5e2ef224cc8e717b20e4910216139
where the SPH kernel length was changed. We changed the normalisation of
the SPH kernel, but omitted to change it for the optical depth. Solve
this by normalising inside sph_kernel instead, so it is only in one
place. Also adjust the tests.