Another attempt to make it build on mac with homebrew or anaconda.
It seems that just finding the directory of the dynamic library and
adding it with -L does not always work. I must be missing something here
because I feel that this ought to work...

One difficulty is that because python is usually
installed from some third party on mac there seems to be a lot of
different failure modes, and I can't test all of them.

THEREFORE tell the linker every way we know how which python library to
link, both with -L and with a -Wl,-L to pass it directly to the linker.
Furthermore, make the python import a weak symbol, in the hope that when
it is resolved at runtime python will already be loaded and it will pick
up the right symbol from the runtime environment it finds itself in.

Hopefully this will fix the actual case we are interested in.