Patch 5 (#26)
Some minor changes to add:
- random noise calculation : In observations, this random noise is different for each sightline (different observing time for each spectra, I guess), so instead of a single snr in the argument of spectra() the code needs an array.
- Continuum fitting error : I just used the continuum error modelling which Drew has used in his observations.

I tired to keep the code general to be used for any observation. For instance, snr and CE are the random variables should be modelled from observation, and then to be passed to fake_spectra. If no noise is needed, the default values will not add any to the spectra.