no need to open snapshot when reading gridded spectra from files (#56)
Until now GriddedSpectra always needed access to the snapshot even when reading pre-computed spectra from file, since it needs BoxSize to compute the grid positions of the lines of sight.

However, this information is already contained in written files, so no need to compute these when reading from file. This is important since often the snapshots are very large and might not be available at the time of reading (I store them on tape).

This functionality was already present in Spectra, it is only GriddedSpectra that needs this small update.